git-svn-id: https://iiop-net.svn.sourceforge.net/svnroot/iiop-net/trunk@433 b113d865-ab28-0410-9ae9-d2690f63996d

+++ IIOPNet/IIOPChannel/GiopMessageBodySerializer.cs
@@ -248,7 +248,7 @@ namespace Ch.Elca.Iiop.MessageHandling {
             }
             
             // for GIOP >= 1.2, a union is used for target information
-            ulong targetAdrType = cdrStream.ReadULong();
+            uint targetAdrType = cdrStream.ReadULong();
             switch (targetAdrType) {
                 case 0:
                     return ReadTargetKey(cdrStream);
